requirements.txt
amqplib==1.0.2
attrs==19.3.0
Automat==0.8.0
bitarray==1.1.0
certifi==2019.9.11
chardet==3.0.4
constantly==15.1.0
Cython==0.29.14
graypy==2.1.0
hyperlink==19.0.0
idna==2.8
impyla==0.16.2
incremental==17.5.0
mysql-connector-python==8.0.18
pika==1.1.0
ply==3.11
protobuf==3.10.0
pure-sasl==0.6.2
PyHamcrest==1.9.0
requests==2.22.0
requests_html==0.10.0
retrying==1.3.3
sasl==0.2.1
selenium==3.141.0
six==1.13.0
thrift==0.13.0
thrift-sasl==0.2.1
thriftpy==0.3.9
thriftpy2==0.4.8
Twisted==19.10.0
urllib3==1.25.7
zope.interface==4.7.1
# pandas==1.0.3
# python-dateutil==2.8.1
# pytz=2019.3
# et-xmlfile=1.0.1
# jdcal=1.4.1
# openpyxl=3.0.3

